requirements.txt
# Основные зависимости для доменной модели
sqlalchemy>=1.4.0
alembic>=1.7.0

# FastAPI и связанные зависимости
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
pydantic>=1.8.0

# Дополнительные утилиты
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# HTTP клиенты
httpx>=0.19.0
requests>=2.28.0

# Работа с переменными окружения
python-dotenv>=1.0.0

# AI/LLM интеграции
openai>=1.0.0

# Для разработки и тестирования
pytest>=6.2.0
pytest-asyncio>=0.15.0

# Для работы с базой данных (опционально)
psycopg2-binary>=2.9.0  # Для PostgreSQL
pymysql>=1.0.0  # Для MySQL
